requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=7_kmp_llvm
accelerate=1.12.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.13.3=pypi_0
aiosignal=1.4.0=pypi_0
alembic=1.18.4=pypi_0
annotated-doc=0.0.4=pypi_0
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
anyio=4.12.1=pypi_0
asteroid-filterbanks=0.4.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.4.0=pypi_0
av=16.1.0=pypi_0
backports.zstd=1.3.0=py310h69bd2ac_0
bitsandbytes=0.49.2=pypi_0
blas=1.0=mkl
blinker=1.9.0=py310h06a4308_1
brotli-python=1.2.0=py310hba01987_1
bzip2=1.0.8=h5eee18b_6
ca-certificates=2026.1.4=hbd8a1cb_0
certifi=2026.01.04=py310h06a4308_0
cffi=2.0.0=pypi_0
charset-normalizer=3.4.4=pyhd8ed1ab_0
click=8.3.1=pypi_0
coloredlogs=15.0.1=pypi_0
colorlog=6.10.1=pypi_0
contourpy=1.3.2=pypi_0
cpython=3.10.19=py310hd8ed1ab_3
ctranslate2=4.7.1=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.9.19=0
cuda-runtime=12.1.0=0
cuda-version=12.9=3
cycler=0.12.1=pypi_0
einops=0.8.2=pypi_0
exceptiongroup=1.3.1=pypi_0
expat=2.7.4=h7354ed3_0
faster-whisper=1.2.1=pypi_0
ffmpeg=4.4.0=hca11adc_2
ffmpeg-python=0.2.0=pypi_0
filelock=3.23.0=pyhd8ed1ab_0
flask=3.1.2=py310h06a4308_0
flatbuffers=25.12.19=pypi_0
fonttools=4.61.1=pypi_0
freetype=2.14.1=ha770c72_0
frozenlist=1.8.0=pypi_0
fsspec=2026.2.0=pypi_0
future=1.0.0=pypi_0
giflib=5.2.2=hd590300_0
gmp=6.3.0=hac33072_2
gmpy2=2.2.1=py310h63ebcad_2
gnutls=3.6.13=h85f3911_1
greenlet=3.3.1=pypi_0
h11=0.16.0=pypi_0
h2=4.3.0=pyhcf101f3_0
hf-xet=1.2.0=pypi_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.36.2=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.1.0=pyhd8ed1ab_0
hyperpyyaml=1.2.3=pypi_0
icu=78.2=h33c6efd_0
idna=3.11=pyhd8ed1ab_0
intel-openmp=2025.0.0=h06a4308_1171
itsdangerous=2.2.0=py310h06a4308_0
jinja2=3.1.6=pyhcf101f3_1
joblib=1.5.3=pypi_0
julius=0.2.7=pypi_0
kiwisolver=1.4.9=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.18=h0c24ade_0
ld_impl_linux-64=2.44=h153f514_2
lerc=4.0.0=h0aef613_1
libblas=3.9.0=20_linux64_mkl
libcblas=3.9.0=20_linux64_mkl
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.14.1.1=4
libcurand=10.3.10.19=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.25=h17f619e_0
libexpat=2.7.4=h7354ed3_0
libffi=3.4.4=h6a678d5_1
libfreetype=2.14.1=ha770c72_0
libfreetype6=2.14.1=h73754d4_0
libgcc=15.2.0=h69a1729_7
libgcc-ng=15.2.0=h166f726_7
libgomp=15.2.0=h4751f2c_7
libhwloc=2.12.2=default_hafda6a7_1000
libiconv=1.18=h3b78370_2
libjpeg-turbo=3.1.2=hb03c661_0
liblapack=3.9.0=20_linux64_mkl
liblzma=5.8.2=hb03c661_0
liblzma-devel=5.8.2=hb03c661_0
libnpp=12.0.2.50=0
libnsl=2.0.0=h5eee18b_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.55=h421ea60_0
libstdcxx=15.2.0=h39759b7_7
libstdcxx-ng=15.2.0=hc03a8fd_7
libtiff=4.7.1=h9d88235_1
libuuid=1.41.5=h5eee18b_0
libvpx=1.11.0=h9c3ff4c_3
libwebp=1.6.0=h9635ea4_0
libwebp-base=1.6.0=hd42ef1d_0
libxcb=1.17.0=h9b100fa_0
libxml2=2.15.1=he237659_1
libxml2-16=2.15.1=hca6bf5a_1
libzlib=1.3.1=hb25bd0a_0
lightning=2.6.1=pypi_0
lightning-utilities=0.15.2=pypi_0
llvm-openmp=15.0.7=h0cdce71_0
mako=1.3.10=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=py310h3406613_0
matplotlib=3.10.8=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.2.0=ha770c72_50498
mkl-include=2025.3.0=hf2ce2f3_463
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
multidict=6.7.1=pypi_0
ncurses=6.5=h7934f7d_0
nettle=3.6=he412f7d_0
networkx=3.4.2=pyh267e887_2
numpy=1.26.4=pypi_0
ocl-icd=2.3.3=hb9d3cd8_0
omegaconf=2.3.0=pyhd8ed1ab_0
onnxruntime=1.23.2=pypi_0
opencl-headers=2025.06.13=h5888daf_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.4=h55fea9a_0
openssl=3.6.1=h35e630c_1
optuna=4.7.0=pypi_0
packaging=25.0=py310h06a4308_1
pandas=2.3.3=pypi_0
pillow=12.1.1=py310h5a73078_0
pip=26.0.1=pyhc872135_0
primepy=1.3=pypi_0
propcache=0.4.1=pypi_0
protobuf=6.33.5=pypi_0
psutil=7.2.2=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pyannote-audio=3.1.1=pypi_0
pyannote-core=6.0.1=pypi_0
pyannote-database=6.1.1=pypi_0
pyannote-metrics=4.0.0=pypi_0
pyannote-pipeline=4.0.0=pypi_0
pycparser=3.0=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.3.2=pypi_0
pysocks=1.7.1=pyha55dd90_7
python=3.10.19=h6fa692b_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.2.1=pypi_0
python_abi=3.10=2_cp310
pytorch=2.5.1=py3.10_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_6
pytorch-lightning=2.6.1=pypi_0
pytorch-metric-learning=2.9.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pyyaml=6.0.3=py310h3406613_1
readline=8.3=hc2a1206_0
regex=2026.1.15=pypi_0
requests=2.32.5=pyhcf101f3_1
rich=14.3.2=pypi_0
ruamel-yaml=0.18.17=pypi_0
ruamel-yaml-clib=0.2.15=pypi_0
safetensors=0.7.0=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
semver=3.0.4=pypi_0
sentencepiece=0.2.1=pypi_0
setuptools=80.10.2=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sounddevice=0.5.5=pypi_0
soundfile=0.13.1=pypi_0
speechbrain=1.0.3=pypi_0
sqlalchemy=2.0.46=pypi_0
sqlite=3.51.1=he0a8d7e_0
sympy=1.13.1=pypi_0
tbb=2021.13.0=hb700be7_5
tensorboardx=2.6.4=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.20.3=pypi_0
tomli=2.4.0=pypi_0
torch-audiomentations=0.12.0=pypi_0
torch-pitch-shift=1.2.5=pypi_0
torchaudio=2.5.1=py310_cu121
torchmetrics=1.8.2=pypi_0
torchtriton=3.1.0=py310
torchvision=0.20.1=py310_cu121
tqdm=4.67.3=pypi_0
transformers=4.45.2=pypi_0
typer=0.23.1=pypi_0
typer-slim=0.23.1=pypi_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025.3=pypi_0
urllib3=2.6.3=pyhd8ed1ab_0
werkzeug=3.1.3=py310h06a4308_0
wheel=0.46.3=py310h06a4308_0
x264=1!161.3030=h7f98852_1
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.8.2=ha02ee65_0
xz-gpl-tools=5.8.2=ha02ee65_0
xz-tools=5.8.2=hb03c661_0
yaml=0.2.5=h280c20c_3
yarl=1.22.0=pypi_0
zlib=1.3.1=hb25bd0a_0
zlib-ng=2.3.3=hceb46e0_1
zstd=1.5.7=hb78ec9c_6
